WIP: Client: Add support for custom shells
Currently, QtWayland allows using custom shell plugins, but only
globally. That's it, all windows must use the same shell integration
plugin.

On the other hand, it goes against wayland ideology where the behavior
of wl_surface is tweaked by its associated surface role and therefore
it's inevitable to use multiple shell surface protocols in a single
application.

This change introduces basic infrastructure to support running Qt
applications with multiple shell surface protocols, for example
xdg-shell and layer-shell-v1.

The main idea behind the proposed API is to let user code change the
shell integration object associated with the QWaylandWindow before it's
mapped on the screen.

When a shell integration is set for the QWaylandWindow, it will create
the corresponding QWaylandShellSurface object.

User code can provide a stable API that accesses the created
QWaylandShellSurface object under the hood to initialize the shell
surface. That's it, some of the custom shell will have to use unstable
API. It's recommended to split the implementation of custom shells in
two halves - one provides the user a way to change the state of the
shell surface, the other interacts with QWaylandWindow and possibly
other low level bits.

Note that a QWaylandShellSurface must create its surface role object in
QWaylandShellSurface::create(), which is going to be called when the
window is shown.

When the window is hidden, QWaylandShellSurface::destroy() is going to
be called and the shell surface must reset its state.

QWaylandShellSurface::create() and QWaylandShellSurface::destroy() can
be called multiple times throughout the lifetime of a window.

QWaylandDisplay::shellIntegration() indicates the default shell.

Change-Id: Ie9068a8b0b3082d448575412da35722dff87db1d