Client: Let custom QWaylandShellIntegration associate private data with QWaylandWindow
A QWaylandShellIntegration may need to associate its private data with a
QWaylandWindow, for example to set the initial state of the window, before
the shell surface is created.

This change introduces QWaylandShellData that can be used by shell
integration plugins to store their private data associated with
QWaylandWindow. QWaylandShellData is owned by the QWaylandWindow.

Change-Id: I30c506f9c75fe0628930a62ab137e5a86e256c5a