Client: Introduce a hook to attach popups to parent shell surface
xdg_popup can have a parent surface that's not an xdg_surface, for
example a layer_surface.

If that happens, you need to pass nil parent surface to
xdg_surface.get_popup request and use layer_surface.get_popup request to
establish parent-child connection between the surfaces.

QWaylandShellSurface::attachPopup and QWaylandShellSurface::detachPopup
provides a way to create a backlink.

A few words about the API design. We could add new signals in
QWaylandWindow, but the problem is the timing. The get_popup request
should be called before the first surface commit, but it makes sense to
emit a childPopupAdded signal after finishing initializing the popup.
Also, it would require some significant changes in the existing code as
QWaylandWindow::shellSurface() would not return a valid shell surface by
that time yet and therefore we won't be able to inspect its
surfaceRole() to get the xdg_popup object.

With the introduction of an xdg positioner type, we could get rid of
popup child tracking code in QWaylandShellSurface and leave only the
hooks to setup/break parent-child connection.

Change-Id: Ice70577d325e6a4a4dc752cab5b090f41cac6d96