Client: Resize wl_egl_window when starting painting
Currently wl_egl_window calls are made from both the main and render
thread for the same window, this is not explicitly allowed in the
specification and breaks the Nvidia driver. In addition it's possible to
call updateSurface from both the main and render thread concurrently
which is not handled properly. This patch moves all EGL surface handling
into being consistently within the render thread.

Fixes: QTBUG-95817
Change-Id: Ic987880c3d4a1220044fb4e8b3eeac0a4e8c06ea