Client: Drop QWaylandBuffer::setDeleteOnRelease()
Some compositors are known to treat deleting pending buffer as attaching
a null buffer, but if a buffer is destroyed after it is commited, the
expected behavior is well defined, so the code can be simplified.

Also, if QWaylandWindow::safeCommit() doesn't commit the surface, the
release event will never be sent and the graphics buffer memory will be
leaked.

Change-Id: Ie01aed7887355f799f84d14d37bc2337e8b9a914