Client: Process frame callbacks and wl_buffer.release in same event queue
Frame callbacks are dispatched from the frame event queue. On the other
hand, the wl_buffer.release events will be processed when dispatching
the default event queue.

Because of different event queue affinity, some QWaylandBuffer objects
can be viewed as busy after receiving a frame callback even though they
are not anymore, the default event queue has not been dispatched yet.

This moves frame callbacks and shm buffers to the same event queue to
ensure consistent buffer state. Note that it relies on the compositor to
send wl_buffer.release events before posting frame callbacks, which is
the case with most of the mainstream compositors.

Change-Id: Ieded4ab4ab7ea542ab138abc6c736ce5add4ce40